pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aqi-prediction-mlops"
version = "1.0.0"
description = "Production-grade AQI prediction system for India with MLOps pipeline"
authors = [{name = "bhautik vekariya", email = "bhautikvekariya1123@gmail.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9,<3.12"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "numpy>=1.24.0,<2.0.0",
    "pandas>=2.0.0,<3.0.0",
    "scikit-learn>=1.3.0,<2.0.0",
    "xgboost>=2.0.0,<3.0.0",
    "catboost>=1.2.0,<2.0.0",
    "fastapi>=0.103.0,<1.0.0",
    "uvicorn[standard]>=0.23.0,<1.0.0",
    "mlflow>=2.7.0,<3.0.0",
    "dvc>=3.27.0,<4.0.0",
    "PyYAML>=6.0.0,<7.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "flake8>=6.1.0",
    "jupyter>=1.0.0",
    "plotly>=5.17.0",
]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  | \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=html --cov-report=term"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true